requirements-azureBlob.txt
flask
pycryptodome
pinecone[grpc]==5.4.2
boto3
aiohttp
requests
asyncio
gunicorn
PyMuPDF
pandas
openpyxl
pyexcel 
pyexcel-ods
google-auth
google-auth-oauthlib
google-auth-httplib2
redis 
Flask-Limiter[redis]
flask_httpauth
google-cloud-pubsub
google-cloud-storage
google-cloud-aiplatform
google-cloud-firestore
openai
pyarmor
cython
psycopg2-binary
tiktoken
python-dateutil
azure-storage-blob
azure-identity